왜 UART DMA가 안되는가? ㅎ

+++ firmware/src/main/msp/fc_msp.c
@@ -438,9 +438,9 @@ static bool mspFcProcessOutCommand(uint16_t cmdMSP, sbuf_t *dst, mspPostProcessF
 
           sbufWriteU16(dst, (getEstimatedAltitudeCm()*10));
 
-          sbufWriteU16(dst, (rcCommand[ROLL]*100));
-          sbufWriteU16(dst, (rcCommand[PITCH]*100));
-          sbufWriteU16(dst, (rcCommand[YAW]*10));
+          sbufWriteU16(dst, (int16_t)(rcCommand[ROLL]*100));
+          sbufWriteU16(dst, (int16_t)(rcCommand[PITCH]*100));
+          sbufWriteU16(dst, (int16_t)(rcCommand[YAW]*10));
           sbufWriteU16(dst, (rcData[THROTTLE]*10));
 
           sbufWriteU32(dst, posllh.lat);
